% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.r
\name{get_endpoint_data}
\alias{get_endpoint_data}
\title{Retrieve detailed endpoint information}
\usage{
get_endpoint_data(host, ip, from_cache = "off")
}
\arguments{
\item{host}{hostname; required.}

\item{ip}{endpoint IP address}

\item{from_cache}{always deliver cached assessment reports if available; optional,
defaults to "off". This parameter is intended for API consumers that
don't want to wait for assessment results. Can't be used at the same
time as the start_new parameter.}
}
\description{
This call is used to retrieve detailed endpoint information. It will return a single
Endpoint object on success. The object will contain complete assessment information.
This API call does not initiate new assessments, even when a cached report is not
found.
}
\references{
\url{https://github.com/ssllabs/ssllabs-scan/blob/stable/ssllabs-api-docs.md}
}

