% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.r
\name{get_root_certs}
\alias{get_root_certs}
\title{Retrieve root certificates}
\usage{
get_root_certs(trust_store = c("moz", "apple", "android", "java", "ms"))
}
\arguments{
\item{trust_store}{one of "\code{moz}" (Mozilla), "\code{apple}", "\code{android}", "\code{java}", "\code{ms}" (Windows)}
}
\value{
\code{list} of \code{openssl}-parsed certificates
}
\description{
This call returns the latest root certificates(Mozilla, Apple MacOS, Android,
Java and Windows) used for trust validation
}
\examples{
get_root_certs()
}
\references{
\url{https://github.com/ssllabs/ssllabs-scan/blob/stable/ssllabs-api-docs-v3.md}
}
