% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.r
\name{get_root_certs_raw}
\alias{get_root_certs_raw}
\title{Retrieve root certificates}
\usage{
get_root_certs_raw()
}
\description{
This call returns the root certificates used for trust validation.
}
\references{
\url{https://github.com/ssllabs/ssllabs-scan/blob/stable/ssllabs-api-docs.md}
}

