% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.r
\name{get_status_codes}
\alias{get_status_codes}
\title{Retrieve known status codes}
\usage{
get_status_codes()
}
\description{
This call will return one StatusCodes instance.
}
\references{
\url{https://github.com/ssllabs/ssllabs-scan/blob/stable/ssllabs-api-docs.md}
}

