% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-info.r
\name{ssllabs_api_info}
\alias{ssllabs_api_info}
\title{Check SSL Labs availability}
\usage{
ssllabs_api_info()
}
\description{
This call should be used to check the availability of the SSL Labs servers, retrieve
the engine and criteria version, and initialize the maximum number of concurrent
assessments. Returns one Info object on success.
}
\examples{
ssllabs_api_info()
}
\references{
\url{https://github.com/ssllabs/ssllabs-scan/blob/stable/ssllabs-api-docs-v3.md}
}
